primitive: number: GenericInt support
Prior to this change, Numbers parsed from a str were created as an
I64, and consequently a lot of downstream usage was tied to i64/I64.
This was particularly manifest during deserialization; despite a
string not tying numeric literals to any particular type, we'd lock in
the i64 type during tokenization and not be able to cleanly cast to a
different type from there.

This change adds a GenericInt variant, which can be created either
through Number::generic or by parsing from a string. All integer types
can be try_from'd with a GenericInt, can be compared with a
GenericInt, and operations b/w generic and specific int types will
coerce the generic int if possible.

Note that rust coerces unconstrained numeric literals to i32, so
Number::from(2) will be an I32, not GenericInt. Number::generic(2)
will side-step that default behavior.

+++ tests/base_serde_test.rs
@@ -62,17 +62,17 @@ fn test_struct() {
 fn test_enum() {
     #[derive(Debug, Deserialize, PartialEq, Serialize)]
     struct Sub {
-        i: i64,
-        am: i64,
-        groot: i64,
+        i: i32,
+        am: u64,
+        groot: i8,
     }
 
     #[derive(Debug, Deserialize, PartialEq, Serialize)]
     enum Test {
         Unit,
-        Int(i64),
+        Int(i8),
         Seq(Vec<f64>),
-        Tuple(i64, i64),
+        Tuple(u16, i32),
         Struct(Sub),
     }
